⚠ Overcomplicate logic for random chains
We want to allow for more chain scenarios, e.g. add a random chain where
both source and target node are in chains themselves. This led to
overcomplicated code having to deal with chains that are placed on edges
instead of nodes. Thus, this commit is a working example for better
chain placement (although not perfect), but definitely needs to be
redone in the next commits (e.g. using the already-used mapping from H
to G instead of essentially redundant chains on edges).