Shift from edge to node supernode encoding
Beforehand, we encoded supernodes using edge costs, e.g. edge with
cost 0 was no chain, an edge with cost 1 indicated that the two nodes
mapped to the same node_H, thus they belong to the same supernode.

However, this was essentially redundant as we also stored the mapping
between input graph H and hardware graph G in the mapping class. We
leverage this class now and encode supernodes directly, thus getting rid
of edge chains (remnants are still there, will clean up later). It might
be worth to keep the term "chain" to indicate supernodes that consist
of more than one node.

This change makes the embedding itself worse, e.g. K4 does not work
anymore. This is since we now also allow to merge random nodes to
supernodes where both source and target are in chains. We need to
implement the evolutionary strategy where we select the mutation
that yields the best result, in our case where the most edges can be
embedded (to achieve a "real" local maximum).

Also changed:
- Changed lists to sets in various places
- Adapted drawing class to the new changes (to properly draw supernodes)
- Made logging output more precise