[Py] Improve embeddings (#12)
* ⚠ Overcomplicate logic for random chains

We want to allow for more chain scenarios, e.g. add a random chain where
both source and target node are in chains themselves. This led to
overcomplicated code having to deal with chains that are placed on edges
instead of nodes. Thus, this commit is a working example for better
chain placement (although not perfect), but definitely needs to be
redone in the next commits (e.g. using the already-used mapping from H
to G instead of essentially redundant chains on edges).

* Rename from_node->source, to_node->target

This only affects the embedding and solver module.

* Shift from edge to node supernode encoding

Beforehand, we encoded supernodes using edge costs, e.g. edge with
cost 0 was no chain, an edge with cost 1 indicated that the two nodes
mapped to the same node_H, thus they belong to the same supernode.

However, this was essentially redundant as we also stored the mapping
between input graph H and hardware graph G in the mapping class. We
leverage this class now and encode supernodes directly, thus getting rid
of edge chains (remnants are still there, will clean up later). It might
be worth to keep the term "chain" to indicate supernodes that consist
of more than one node.

This change makes the embedding itself worse, e.g. K4 does not work
anymore. This is since we now also allow to merge random nodes to
supernodes where both source and target are in chains. We need to
implement the evolutionary strategy where we select the mutation
that yields the best result, in our case where the most edges can be
embedded (to achieve a "real" local maximum).

Also changed:
- Changed lists to sets in various places
- Adapted drawing class to the new changes (to properly draw supernodes)
- Made logging output more precise

* Rewrite extend_random_supernode and fix bugs

Split the method into multiple pieces for better readability and fixed
several bugs that occurred due to wrong order of execution. It is still
quirky that we can't embed edges freely since all nodes on that edge
must have supernode (in the current implementation). That's why there
are some special constructs, e.g. we need to discard nodes from sets
to make sure not the change results after having embedded them
in supernodes. There is room for improvement here.

Fixed a really annoying bug: Sets in Python are returned by reference,
so we got weird results since we used discard on them assuming we just
altered the local variable. Now we copy the set.

Fixed another, potential bug where we accidentally added selfloops to
the embedding view graph.

Also improved error messages.

* Get rid of chain remnants

Note that at this point, we still sometimes use the word "chain",
however we don't use chains encoded on edges anymore, but now on
nodes themselves.

TODO: Unify usage of words "chain" and "supernode" throughout the code.

* Explicitly use sets instead of lists

* Draw supernode colors on nodes (not only on edges)

Also slightly shifted labels of Chimera graph for better readability.

* Change color brightness instead of transparency

* Fix bug not operating on playground embedding

* Fix bug forgot removing target from prev supernode

* Init basic hill climbing (evolution)

After each generation we also try to remove unnecessary edges.

* Move supernode connectiveness check to embedding

* Remove redundant nodes probabilistically

* Implement articulation point algorithm (cut node)

This way, we don't remove the wrong nodes in our graph that would
disconnect the respective subgraph of supernodes (if removed).

See https://www.geeksforgeeks.org/articulation-points-or-cut-vertices-in-a-graph/
and https://youtu.be/jFZsDDB0-vo as examples

* Fix removal of redundant nodes

Previously, we removed all nodes that were not articulation points
for every supernode in one go. However, as we remove, the articulation
points might change, so we need to recalculate them every time a node
is removed.

* Add strategy to extend supernode without shifting

We don't shift the target node in this strategy.

* Fix articulation point (don't include removed nodes)

* Remove redundant nodes as last resort in the end

Also switched to 5x5 Chimera grid

At this point K6 embedding on a 3x3 cell is working. Tried K12 on a
5x5 Chimera grid, which is not yet yielding good valid embeddings.
TODO: Code cleanup, especially the embedding solver that got
kind of cluttered.