Establish "extend supernode without shift" as own mutation
Beforehand, this mutation was only a surrogate if the other strategy
didn't work. We now establish it as full-value mutation and switch
between the two mutations based on a random value.

This also means that we now make sure in the shifting/bumping
mutation that source and target are not from the same supernode
since this would undermine the other strategy which is essentially this.